[Tiny] The parameter name FN is changed to OP.
I call all of function, macro and special operators as ``operators'' in
one piece. I'm now planning to modify the PAPPLY macro series to accept
other than functions as its first parameter.

Actually speaking, it was planned when I wrote APAPPLY macro and
therefore the parameter that takes a function is named OP in APAPPLY
macro.