P macro is fixed to use form format.
Unlike to the old enumerate format, the newly introduced function format
only handles functions. The enumerate format checked if its first
argument was a symbol of operator, not limited to functions but
including macros and special operators, and handled them properly.
However, this powerful feature is intentionally thrown away. This
limitation is one of the main purpose of the modification this time.
The enumerate format and form format has the same power and thus just
adopting the form format instead fix the issue.